x86/resctrl: Wrap resctrl_arch_find_domain() around rdt_find_domain()
rdt_find_domain() finds a domain given a resource and a cache-id.
Its not quite right for the resctrl arch API as it also returns the
position to insert a new domain.

Wrap rdt_find_domain() in a another function resctrl_arch_find_domain()
so we avoid the unnecessary argument outside the arch code.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>