x86/resctrl: Move max_{name,data}_width into resctrl code
max_name_width and max_data_width are used to pad the strings in the
resctrl schemata file.

This should be part of the fs code as it influences the user-space
interface, but currently max_data_width is generated by the arch init code.

Move the variables and max_data_width's initialisation code to rdtgroup.c.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>