x86/resctrl: Create helper for RMID allocation and mondata dir creation
When monitoring is support, each monitor and control group is allocated
an RMID. For control groups, rdtgroup_mkdir_ctrl_mon() later goes on to
allocate the CLOSID.

MPAM's equivalent of RMID are not an independent number, so can't be
allocated until the CLOSID is known. An RMID allocation for one CLOSID
may fail, whereas another may succeed depending on how many monitor
groups a control group has.

The RMID allocation needs to move to be after the CLOSID has been
allocated.

Move the RMID allocation and mondata dir creation to a helper, this
makes a subsequent change easier to read.

Tested-by: Shaopeng Tan <tan.shaopeng@fujitsu.com>
Reviewed-by: Ilpo Järvinen <ilpo.jarvinen@linux.intel.com>
Signed-off-by: James Morse <james.morse@arm.com>
---
Changes since v4:
 * Fixed typo in commit message, moved some words around.
Signed-off-by: Tushar Dave <tdave@nvidia.com>