arm_mpam: Assign msc id early into driver probe
msc id is widely used within the mpam driver as an aid for error
logging. Since the id is being assigned towards the end of the driver
probe, the error logs captured before this assignment reports zero as
the id irrespective of what the actual id for the msc is. Resolve this
issue by moving the msc id assignment early into the driver probe.

Signed-off-by: Rohit Mathew <rohit.mathew@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>