arm_mpam: Fix ID allocation upper limit for monitors
ida framework is used to allocate and free CSU/MBWU monitors as and when
they are requested. "ida_alloc_range" is used to get the next free ID
from a range of IDs that are not in use. The function takes the minimum
and maximum ranges (both inclusive) so that a free ID could be allocated
from the range. The value returned is, at a later point programmed into
the MSMON_CFG_MON_SEL register. Both CSU and MBWU monitors range from 0
to NUM_MON - 1. As of now, the "ida_alloc_range" could very well return
the upper limit NUM_MON for both CSU and MBWU monitors which could lead
to a out of bound monitor access.

Fix this issue by allocating monitors for both CSU and MBWU from 0 to
NUM_MON - 1.

Signed-off-by: Rohit Mathew <rohit.mathew@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>