ACPI/irq: Apply ACPI_IRQCHIP_FWSPEC_ARG0 only for GSI domain
The implementation of the updated GICv3 parameter parsing code in
pack_fwspec() is not compatible for the non-GSI drivers. It uses
the offset 1 for hardware IRQ, but the unmodified GPIO driver
expects the offset value to be 0, leading to a regression.

This patch applies new mapping to the GSI driver only to fix
the issue.

Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>