drivers/perf: Fix kernel panic due to the invalid mon_ctx pointer
The return pointer from the resctrl_arch_mon_ctx_alloc_no_wait() function
is saved in a 32-bit variable 'hwc->idx' which results in the loss of
the upper 32 bits. This, in turn, triggers a kernel panic when attempting
to access a corrupted pointer.

Use 'event->pmu_private' instead of 'hwc->idx' to resolve the issue.

Another reason contributing to the issue due to continuous 'error irq'
messages. When the resctrl_arch_mon_ctx_alloc_no_wait() function is unable
to allocate a monitor resource, returns a valid PTR, this results in the
incorrect programming of MON_SEL. Return a ERR_PTR if failed to allocate
monitor.

dmesg:
  mpam: error irq from msc:0 'Monitor_Range', partid:228, pmg: 0, ris: 0
  mpam: error irq from msc:0 'Monitor_Range', partid:228, pmg: 0, ris: 0
  mpam: error irq from msc:0 'Monitor_Range', partid:228, pmg: 0, ris: 0
  ...
  mpam: error irq from msc:0 'Monitor_Range', partid:228, pmg: 0, ris: 0

Change-Id: Ib77cd93c40f9359245faa81992fef07e23dca9f7
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>