arm_mpam: resctrl: Update only requested configuration
The resctrl has helper functions for updating CPOR and MBW configuration
updates. While the existing resctrl_arch_update_one() effectively updates
the specified configuration as intended, it inadvertently overrides other
configurations with default values.

Example bandwdith value is not applied:
 root# cat /sys/fs/resctrl/schemata
 MB:1=100
 L3:1=fff

 root# echo -e "L3:1=fff\nMB:1=50" > /sys/fs/resctrl/schemata
 root# cat /sys/fs/resctrl/schemata
 MB:1=100
 L3:1=fff

Fix the potential loss of accuracy during the conversion of MBW_MAX from
percentage to fixed-point representation, and vice versa. The updated
functions provide fixed-point values that closely align with the values
specified in the MPAM specification, Table 9-3 for Fraction Widths And
Hex Representation.

Before this fix:
 root# echo -e "MB:1=1" > /sys/fs/resctrl/schemata
 root# cat /sys/fs/resctrl/schemata
 MB:1=000
 L3:1=fff

 root# echo -e "MB:1=2" > /sys/fs/resctrl/schemata
 root# cat /sys/fs/resctrl/schemata
 MB:1=001
 L3:1=fff

 root# echo -e "MB:1=3" > /sys/fs/resctrl/schemata
 root@pset# cat /sys/fs/resctrl/schemata
 MB:1=001
 L3:1=fff

With this patch:
 root# echo -e "MB:1=1" > /sys/fs/resctrl/schemata
 root# cat /sys/fs/resctrl/schemata
 MB:1=001
 L3:1=fff

 root# echo -e "MB:1=2" > /sys/fs/resctrl/schemata
 root# cat /sys/fs/resctrl/schemata
 MB:1=002
 L3:1=fff

 root# echo -e "MB:1=3" > /sys/fs/resctrl/schemata
 root@pset# cat /sys/fs/resctrl/schemata
 MB:1=003
 L3:1=fff

Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>
[tdave: updated patch and commit description ]
Link: https://patchwork.kernel.org/project/linux-arm-kernel/patch/20240819220752.201815-1-sdonthineni@nvidia.com/
Signed-off-by: Tushar Dave <tdave@nvidia.com>