iommu/arm-smmu-v3: Make arm_smmu_tlb_inv_range_s2()
This is the mirror of arm_smmu_tlb_inv_range_s1(), it is made by taking
the S1 functionality out of arm_smmu_tlb_inv_range_domain() and renaming
it.

This pushes the decision of s1/s2 up directly to functions which are
function pointers setting the stage to just have s1/s2 specific ops and
remove the s1/s2 test entirely.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>